feat: update calculations for layering (effective SP by biggest protector, degradation by top layer), add protected by to history, improve folder structure
