+++ src/pages/armor.astro
@@ -108,5 +108,5 @@ import '../styles/armor.css';
 
     </div>
   </Layout>
-  <script src="../scripts/armor/index.ts"></script>
+  <script src="../scripts/armor/render/index.ts"></script>
 </HTML>
